requirements.txt
et-xmlfile==1.1.0
numpy==1.21.6
openpyxl==3.1.3
pandas==1.3.5
Pillow==9.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
